\name{HTMLcolours}
\docType{data}
\alias{HTMLcolours}
\alias{HTMLcolors}
\title{
  HTML colour names
}
\description{
  The list of HTML colour names 
}
\usage{
HTMLcolours
}
\format{ A \code{"colourList"} object. }
\references{
  https://www.w3.org/TR/2011/REC-css3-color-20110607/#html4
}
\keyword{ datasets }% use one of  RShowDoc("KEYWORDS")
