\name{NgaTae}
\docType{data}
\alias{NgaTae}
\title{
  Simple Maori Colour Names
}
\description{
  A small set of Maori colour names.
}
\usage{
NgaTae
}
\format{ A \code{"colourList"} object. }
\references{
  \url{http://www.maori.org.nz/kotereo/default.php?pid=sp148&parent=115}
}
\keyword{ datasets }% use one of  RShowDoc("KEYWORDS")
