\name{roloc-package}
\alias{roloc-package}
\alias{roloc}
\docType{package}
\title{
\packageTitle{roloc}
}
\description{
\packageDescription{roloc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{roloc}
\packageIndices{roloc}

The main function in this package is \code{colourName()}, which takes
one or more colour specifications and returns a matching colour name.
}
\author{
\packageAuthor{roloc}

Maintainer: \packageMaintainer{roloc}
}
\references{
Murrell, P. (2018). "Generating Colour Names: The 'roloc' Package for R."
Technical Report 2018-01, Department of Statistics, The University of
Auckland.
\url{https://www.stat.auckland.ac.nz/~paul/Reports/roloc/intro/roloc.html}
}
\keyword{ package }
\seealso{
  \code{\link{colourName}}
}
\examples{
colourName("#FF0000")
}
