\name{simpleColours}
\docType{data}
\alias{simpleColours}
\alias{simpleColors}
\title{
  Simple Colour Names
}
\description{
  A set of colour names with descriptive labels that are evenly
  spaced throughout CIE LUV space.
}
\usage{
simpleColours
}
\format{ A \code{"colourList"} object. }
\keyword{ datasets }% use one of  RShowDoc("KEYWORDS")
