\name{Rcolours}
\docType{data}
\alias{Rcolours}
\alias{Rcolors}
\alias{X11colours}
\alias{X11colors}
\alias{CSScolours}
\alias{CSScolors}
\alias{SVGcolours}
\alias{SVGcolors}
\title{
  The standard R colour names
}
\description{
  The list of colour names that R recognises as
  colour specifications.  This list can also be referred to
  as \code{X11colours} or \code{CSScolours} or \code{SVGcolours}.
}
\usage{
Rcolours
}
\format{ A \code{"colourList"} object. }
\references{
Jaffer, A. (2018). Color-name
dictionaries. http://people.csail.mit.edu/jaffer/Color/Dictionaries#X11. Accessed: 2018-01-24.
}
\keyword{ datasets }% use one of  RShowDoc("KEYWORDS")
