\name{ReseneColours}
\docType{data}
\alias{ReseneColours}
\alias{ReseneColors}
\title{
  Resene colour dictionary
}
\description{
  The Resene list of colour names 
}
\usage{
ReseneColours
}
\format{ A \code{"colourList"} object. }
\source{
  http://people.csail.mit.edu/jaffer/Color/Resene-X11
}
\references{
}
\keyword{ datasets }% use one of  RShowDoc("KEYWORDS")
