\name{roloc-package}
\alias{roloc-package}
\alias{roloc}
\docType{package}
\title{
\packageTitle{roloc}
}
\description{
\packageDescription{roloc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{roloc}
\packageIndices{roloc}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{roloc}

Maintainer: \packageMaintainer{roloc}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
