\name{roloc-package}
\alias{roloc-package}
\alias{roloc}
\docType{package}
\title{
\packageTitle{roloc}
}
\description{
\packageDescription{roloc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{roloc}
\packageIndices{roloc}

The main function in this package is \code{colourName()}, which takes
one or more colour specifications and returns a matching colour name.
}
\author{
\packageAuthor{roloc}

Maintainer: \packageMaintainer{roloc}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
}
\examples{
colourName("#FF0000")
}
